Fix dnsmasq.conf
On some systems, the dnsmasq.conf file does not set '/etc/dnsmasq.d' as a conf-dir, and the service does not run with the `--conf-dir` parameter set. This leads to a situation where cookbook settings are not applied to dnsmasq when it runs.

This fixes that issue by parsing the existing conf file (if anything is set), and overriding the 'conf-dir' setting to what is used by the dhcp and dns recipes. This behavior ensures that users can still specify settings in that file elsewhere while pulling in the config from this cookbook.